dubious advancements, but works till send command

+++ src/main/java/it/angelic/growlroom/controllers/ESPActuatorsController.java
@@ -39,7 +39,7 @@ public class ESPActuatorsController {
 	 * @return
 	 */
 	@PutMapping(value = "/id/{id}", produces = MediaType.APPLICATION_JSON_VALUE, consumes = MediaType.APPLICATION_JSON_VALUE)
-	public int putActuator(@PathVariable String id, @RequestBody Actuator dispositivo) {
+	public long putActuator(@PathVariable String id, @RequestBody Actuator dispositivo) {
 		try {
 			int t2 = Integer.valueOf(id);
 		} catch (NumberFormatException e) {
